tools: rudimentary docx importer
Add a rudimentary importer for Microsoft Word docx files. It uses pandoc
to convert the Word documents to Markdown and extracts the embedded images
into static/img/. It's also fixing the paths and names of the images so we
can generally use it. If an author, date or description is provided the
article meta data is patched to use it. Also a fotogrid section will be
created.

What's left is cleaning the generated Markdown from Word's image size
constraints, I have not found a generally usable way that works without
manual intervention. Also resizing the images is not yet done.

Signed-off-by: Johannes Thumshirn <jth@kernel.org>